Test with 8.8.1 and 8.10.2 on travisci
This patch adds a test with the most recent versions of GHC on
travisci. This version of GHC makes the import of `Control.Monad.Fail`
redundant, as a part of the [`MonadFail`][monadfail] conversion. This
will allow us to test our `MonadFail` patches with versions of base
both during the `MonadFail` conversion and after the `MonadFail`
conversion.

[monadfail]: https://wiki.haskell.org/MonadFail_Proposal