Add `MonadFail` constraints to `Thunk`
The `whnfPr` function in `Data.Comp.Thunk` uses the `fail` function of
the `Monad` class. With the acceptance of the [`MonadFail`
proposal][monadfail], a new, bespoke `MonadFail` class is introduced
in the `Control.Monad.Fail` module that contains a `fail` function. In
GHC 8.8.1, `Monad`’s `fail` function is moved out into `MonadFail`,
which itself is moved into `Prelude`. The result of this series of
changes is that the package no longer compiles on GHC ≥ 8.8.1.

This patch adds a `MonadFail` constraint to `whnfPr` and `nfPr`, which
transitively invokes `whnfPr`. In order to make this patch work for
all supported versions of GHC, we wrap the import of
`Control.Monad.Fail` in a preprocessor directive.

This change fixes issue #29, but in doing so, it exposes the same
problem occurs in the benchmarks, and a similar problem in examples.

[monadfail]: https://wiki.haskell.org/MonadFail_Proposal