Add `MonadFail` constraints to benchmarks
Just as with the `Data.Comp.Thunk` module, the benchmarks no longer
compile on GHC ≥ 8.8.1 with the acceptance of the [`MonadFail`
proposal][monadfail]. This patch adds `MonadFail` constraints to
functions that rely on `Monad`’s old `fail` function, and wraps the
imports in preprocessor directives.

This change fixes the same issue discussed in #29 for benchmarks.

[monadfail]: https://wiki.haskell.org/MonadFail_Proposal